Removed getVersion()
getVersion() caused the plugin not to load at all. I'm not sure if invalid JSON is causing it, but removing getVersion() (for now) fixes the problem with it not loading.

The config.json files also needs to be deleted and reloaded (or just updated), but I'm assuming that's what happens when plugins are updated.